refactor: перенос upgrade-window из shared в widgets для соответствия fsd
