+++ src/app/widgets/upgrade-window/upgrade-window.component.css
